% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2-class.R
\name{combineMSMSSpectra}
\alias{combineMSMSSpectra}
\title{COmbines multiples MS-MS spectra as a signle consensus spectrum.}
\usage{
combineMSMSSpectra(lspec, ppmMS = 3, ppmMSMS = 5, SNR = 2,
  smoothSize = 10, type = "FIA", ...)
}
\arguments{
\item{fname}{Tthe mzML file to be parsed.}

\item{mz}{the mass of the ocmpoudns to be found.}

\item{ppm}{the ppm parameter used by the density function.}

\item{msWindows}{the size of the ms windows to search the prcursor in.}

\item{multiplicity}{the multiplicituy to be used for th efiltering of the peaks.}
}
\value{
A list of spectra classified by energy.
}
\description{
Return the consensus spectra of a compounds given a filename and a mass
with one spectra by energy.
}
\examples{
print("examples to be put here")
}

