% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMS2.R
\name{cosineDist}
\alias{cosineDist}
\title{Calculate the cosine similarity.}
\usage{
cosineDist(s1, s2, mzexp = 2, intexp = 0.5, ppm, mzmin = 0.005)
}
\arguments{
\item{s1}{z MSMSpsectrum object}

\item{mzexp}{The wieghting of the mz term.}

\item{intexp}{the weighting of the intensity term.}

\item{ppm}{the tolerance in the matching in ppm}

\item{mzmin}{the minimum tolerance}

\item{s1}{a MSMSpsectrum object}
}
\value{
The cosine similarity between the 2 spectra.
}
\description{
Calculate the cosine similarity btween two MSMSspectrum object.
}
\examples{
print("examples ot be put here")
}

