% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2-class.R
\name{extractMS2spectra}
\alias{extractMS2spectra}
\title{Extract the MS2 spectra form an FIA acquisition.}
\usage{
extractMS2spectra(cfile, ppm = 5, tolMS = 0.5, mzlist = NULL,
  rtlist = NULL, type = c("split", "ordered", "none"),
  paramcentwave = list(ppm = 5, snthresh = 3, peakwidth = c(10, 50), prefilter
  = c(3, 500)))
}
\arguments{
\item{ppm}{the ppm parameter used by the density function.}

\item{fname}{Tthe mzML file to be parsed.}

\item{mz}{the mass of the ocmpoudns to be found.}

\item{msWindows}{the size of the ms windows to search the prcursor in.}

\item{multiplicity}{the multiplicituy to be used for th efiltering of the peaks.}
}
\value{
A list of spectra classified by energy.
}
\description{
Return the consensus spectra of a compounds given a filename and a mass
with one spectra by energy.
}
\examples{
print("examples to be put here")
}

