% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2-class.R
\name{getScanMzr}
\alias{getScanMzr}
\title{Get the MS-MS scan corresponding to a mass.}
\usage{
getScanMzr(msraw, mzprecursor, tol = 1, MSwindows = 10, type = c("split",
  "ordered", "none"), smoothSize = 10)
}
\arguments{
\item{msraw}{A character giving the path to an object or a mzRramp object.}

\item{mzprecursor}{The mass to be looked in.}

\item{tol}{the tolerance un absolute value.}

\item{MSwindows}{The windows in which the MS scan need to be returned.}

\item{type}{Split the list have energy as names. Ordered list is ordered by energy. None as returned by mzR.}

\item{smoothSize}{the size of the smoothing windows in second. Shloud be a third of the expecte peakwidth approximately.}
}
\value{
A list of MSMSspectrum object
}
\description{
All the MS-MS scans corresponding to a ocmpounds are placed in a list of MS2 object.
}
\examples{
print("examples to be put here")
}

