% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2-class.R
\docType{methods}
\name{plot,MSMSspectrum,ANY-method}
\alias{plot,MSMSspectrum,ANY-method}
\title{plot an MSMSspectrum object.}
\usage{
\S4method{plot}{MSMSspectrum,ANY}(x, y = NULL, level = c("MS2", "MS1"),
  xlim = NULL)
}
\arguments{
\item{x}{an MSMS spectrum object}

\item{y}{always null}

\item{level}{the level of plotting "MS1" or "MS2"}

\item{xlim}{an optional windows of mass to be plotted.}
}
\value{
nothing
}
\description{
plot an MSMSspectrum object at the MS-MS level or at the MS level.
}
\examples{
print("examples to be put here")
}

