% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2-class.R
\name{plotPrecursorSeq}
\alias{plotPrecursorSeq}
\title{plot an MSMSspectrum object.}
\usage{
plotPrecursorSeq(lSpec, ylim = NULL, xlim = NULL)
}
\arguments{
\item{xlim}{an optional windows of mass to be plotted.}

\item{x}{an MSMS spectrum object}

\item{y}{always null}

\item{level}{the level of plotting "MS1" or "MS2"}
}
\value{
nothing
}
\description{
plot an MSMSspectrum object at the MS-MS level or at the MS level.
}
\examples{
print("examples to be put here")
}

